[#2285] Remove unused management commands
In case of `check_for_workshop_website_updates`, some methods
were used and had to be extracted out into metadata utils.
